+++ tests/api/sats.test.ts
@@ -63,6 +63,11 @@ describe('/sats', () => {
           transfers_pre_inscription: 0,
           tx_index: 0,
           curse_type: null,
+          inscription_pointer: null,
+          delegate: null,
+          metaprotocol: null,
+          metadata: null,
+          parent: null,
         })
         .build()
     );
@@ -100,6 +105,11 @@ describe('/sats', () => {
           inscription_input_index: 0,
           transfers_pre_inscription: 0,
           tx_index: 0,
+          inscription_pointer: null,
+          delegate: null,
+          metaprotocol: null,
+          metadata: null,
+          parent: null,
         })
         .build()
     );
@@ -132,6 +142,11 @@ describe('/sats', () => {
           inscription_input_index: 0,
           transfers_pre_inscription: 0,
           tx_index: 0,
+          inscription_pointer: null,
+          delegate: null,
+          metaprotocol: null,
+          metadata: null,
+          parent: null,
         })
         .build()
     );