+++ tests/api/stats.test.ts
@@ -224,6 +224,11 @@ function testRevealApply(blockHeight: number, numbers: number[]) {
         transfers_pre_inscription: 0,
         tx_index: 0,
         curse_type: null,
+        inscription_pointer: null,
+        delegate: null,
+        metaprotocol: null,
+        metadata: null,
+        parent: null,
       });
   }
   return block.build();