+++ tests/setup.ts
@@ -1,6 +1,8 @@
 // ts-unused-exports:disable-next-line
 export default (): void => {
-  process.env.CHAINHOOK_NODE_AUTH_TOKEN = 'test';
-  process.env.CHAINHOOK_NODE_RPC_HOST = 'test.chainhooks.com';
+  process.env.ORDHOOK_NODE_AUTH_TOKEN = 'test';
+  process.env.ORDHOOK_NODE_RPC_HOST = 'test.chainhooks.com';
+  process.env.ORDHOOK_NODE_RPC_PORT = '13370';
+  process.env.ORDHOOK_INGESTION_MODE = 'default';
   process.env.PGDATABASE = 'postgres';
 };